pyproject.toml
[project]
name = "mlx-perception-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ftfy>=6.3.1",
    "mlx>=0.30.1",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "regex>=2025.11.3",
]
